refactor: change name to hook.ts
